flatten the plugin_utils directory + address TypedBuilder comments
