+++ apollo-router-core/src/lib.rs
@@ -29,7 +29,7 @@ mod json_ext;
 mod layers;
 mod naive_introspection;
 mod plugin;
-mod plugin_utils;
+pub mod plugin_utils;
 mod query_cache;
 mod query_planner;
 mod request;
@@ -46,7 +46,6 @@ pub use json_ext::*;
 pub use layers::*;
 pub use naive_introspection::*;
 pub use plugin::*;
-pub use plugin_utils::*;
 pub use query_cache::*;
 pub use query_planner::*;
 pub use request::*;
