+++ apollo-router-core/src/plugin_utils/mod.rs
@@ -1,9 +1,7 @@
-#[cfg(feature = "service_mock")]
 pub mod service;
 pub mod structures;
 
-#[cfg(feature = "service_mock")]
 pub use service::{
     MockExecutionService, MockQueryPlanningService, MockRouterService, MockSubgraphService,
 };
-pub use structures::{RouterRequestBuilder, RouterResponseBuilder};
+pub use structures::{RouterRequest, RouterResponse};
