--- apollo-router-core/src/plugin_utils/structures/mod.rs
@@ -1,5 +1,0 @@
-pub mod router_request;
-pub mod router_response;
-
-pub use router_request::RouterRequestBuilder;
-pub use router_response::RouterResponseBuilder;
